feat: restructure fleet pages
