+++ app/javascript/frontend/api/models/Fleet.d.ts
@@ -1,7 +1,9 @@
 type Fleet = {
   id: string
   name: string
+  description: string
   myFleet: boolean
+  publicFleet: boolean
 }
 
 type FleetForm = {
