+++ app/javascript/frontend/core/components/Avatar/index.scss
@@ -57,9 +57,9 @@
   }
 
   &.avatar-large {
-    width: 200px;
-    min-width: 200px;
-    height: 200px;
+    width: 160px;
+    min-width: 160px;
+    height: 160px;
 
     .no-avatar {
       font-size: 60px;
