+++ app/javascript/frontend/core/components/Navigation/FleetNav/index.vue
@@ -7,15 +7,30 @@
       :active="$route.name === 'fleet'"
       exact
     />
+    <NavItem
+      v-if="currentFleet.publicFleet || currentFleet.myFleet"
+      :to="{ name: 'fleet-ships', params: { slug: currentFleet.slug } }"
+      :label="$t('nav.fleets.ships')"
+      :active="shipsNavActive"
+      icon="fad fa-starship"
+    />
     <template v-if="currentFleet.myFleet">
       <NavItem
         :to="{ name: 'fleet-members', params: { slug: currentFleet.slug } }"
         :label="$t('nav.fleets.members')"
+        :active="$route.name === 'fleet-members'"
         icon="fad fa-users"
+      />
+      <NavItem
+        :to="{ name: 'fleet-stats', params: { slug: currentFleet.slug } }"
+        :label="$t('nav.fleets.stats')"
+        :active="$route.name === 'fleet-stats'"
+        icon="fad fa-chart-bar"
       />
       <NavItem
         :to="{ name: 'fleet-settings', params: { slug: currentFleet.slug } }"
         :label="$t('nav.fleets.settings.index')"
+        :active="$route.name === 'fleet-settings'"
         icon="fad fa-cogs"
       />
     </template>
@@ -38,6 +53,10 @@ export default class FleetNav extends Vue {
 
   get currentFleet(): Fleet | null {
     return this.collection.record
+  }
+
+  get shipsNavActive() {
+    return ['fleet-ships', 'fleet-fleetchart'].includes(this.$route.name)
   }
 
   mounted() {
