+++ app/javascript/frontend/plugins/Validations/rules/index.ts
@@ -5,5 +5,15 @@ import usernameTaken from './usernameTaken'
 import user from './user'
 import url from './url'
 import hexColor from './hexColor'
+import text from './text'
 
-export { emailTaken, serialTaken, fidTaken, usernameTaken, user, url, hexColor }
+export {
+  emailTaken,
+  serialTaken,
+  fidTaken,
+  usernameTaken,
+  user,
+  url,
+  hexColor,
+  text,
+}
