+++ app/javascript/frontend/plugins/Validations/rules/text/index.ts
@@ -1,0 +1,9 @@
+const regex: RegExp = /^[0-9A-ZÀÂÆÇÉÈÊËÏÎÔŒÙÛÜŸÄÖßÁÍÑÓÚ\-_'"\.,\?!:;\/\s]*$/i
+
+const validate = (value: string): boolean => regex.test(value)
+
+export { validate }
+
+export default {
+  validate,
+}
