+++ app/javascript/stylesheets/frontend/partials.scss
@@ -8,7 +8,6 @@
 @import 'partials/erkul';
 @import 'partials/filtered-header';
 @import 'partials/filtered-table';
-@import 'partials/fleet';
 @import 'partials/fleetchart';
 @import 'partials/flex-list';
 @import 'partials/form-checkbox';
