+++ app/javascript/stylesheets/frontend/partials/form-input.scss
@@ -6,7 +6,8 @@
   .form-input-wrapper {
     display: flex;
 
-    input {
+    input,
+    textarea {
       display: block;
       width: 100%;
       height: 43px;
@@ -35,6 +36,10 @@
       &:focus {
         outline: none;
       }
+    }
+
+    textarea {
+      height: auto;
     }
 
     .form-input-prefix,
@@ -72,14 +77,16 @@
     }
 
     &.from-input-wrapper-with-prefix {
-      input {
+      input,
+      textarea {
         border-top-left-radius: 0;
         border-bottom-left-radius: 0;
       }
     }
 
     &.from-input-wrapper-with-suffix {
-      input {
+      input,
+      textarea {
         border-top-right-radius: 0;
         border-bottom-right-radius: 0;
       }
@@ -87,13 +94,15 @@
   }
 
   &.form-input-clearable {
-    input {
+    input,
+    textarea {
       padding-right: 50px;
     }
   }
 
   &.form-input-large {
-    input {
+    input,
+    textarea {
       height: 55px;
       padding: 10px 16px;
       font-size: 19px;
@@ -109,14 +118,16 @@
   &.form-input-clean {
     margin-bottom: 0;
 
-    input {
+    input,
+    textarea {
       border: none;
       border-radius: 0;
     }
   }
 
   &.has-error {
-    input {
+    input,
+    textarea {
       border-color: darken($danger, 15%);
       box-shadow: inset 0 1px 1px rgba(0, 0, 0, .075);
     }
@@ -128,13 +139,15 @@
     }
 
     .from-input-wrapper-with-prefix {
-      input {
+      input,
+      textarea {
         border-left: none;
       }
     }
 
     .from-input-wrapper-with-suffix {
-      input {
+      input,
+      textarea {
         border-right: none;
       }
     }
