+++ app/javascript/stylesheets/frontend/partials/icons.scss
@@ -6,24 +6,13 @@
   background-position: center;
   background-size: 100%;
 
-  &.icon-guilded {
-    width: 30px;
-    height: 30px;
-    background-image: url(~images/guilded_logo.png);
-  }
-
   &.icon-rsi {
-    width: 36px;
-    height: 20px;
+    width: 46px;
+    height: 24px;
     background-image: url(~images/rsi_logo.png);
   }
 
   &.icon-label {
-    &.icon-guilded {
-      width: 15px;
-      height: 15px;
-    }
-
     &.icon-rsi {
       width: 30px;
       height: 18px;
@@ -31,11 +20,6 @@
   }
 
   &.icon-small {
-    &.icon-guilded {
-      width: 20px;
-      height: 20px;
-    }
-
     &.icon-rsi {
       width: 22px;
       height: 16px;
@@ -43,11 +27,6 @@
   }
 
   &.icon-large {
-    &.icon-guilded {
-      width: 60px;
-      height: 60px;
-    }
-
     &.icon-rsi {
       width: 76px;
       height: 40px;
