+++ app/javascript/stylesheets/frontend/partials/typography.scss
@@ -7,6 +7,10 @@ h4 {
 h1 {
   margin-top: 0;
   margin-bottom: 15px;
+
+  &.large {
+    font-size: $h1-large-font-size;
+  }
 }
 
 address {
