+++ app/javascript/stylesheets/variables.scss
@@ -61,6 +61,7 @@ $container-max-widths: (
 );
 
 $font-size-base: 1rem;
+$h1-large-font-size:                $font-size-base * 2.5;
 $h1-font-size:                $font-size-base * 2.15;
 $h2-font-size:                $font-size-base * 1.7;
 $h3-font-size:                $font-size-base * 1.25;
