+++ app/views/api/v1/fleet_members/_member.jbuilder
