+++ app/views/api/v1/fleet_members/index.jbuilder
@@ -1,0 +1,3 @@
+# frozen_string_literal: true
+
+json.array! @members, partial: 'api/v1/fleet_members/member', as: :member
