+++ app/views/api/v1/fleet_members/quick_stats.jbuilder
