+++ app/views/api/v1/fleet_vehicles/_model.jbuilder
