+++ app/views/api/v1/fleet_vehicles/fleetchart.jbuilder
