+++ app/views/api/v1/fleet_vehicles/models.jbuilder
@@ -1,0 +1,3 @@
+# frozen_string_literal: true
+
+json.array! @models, partial: 'api/v1/fleet_vehicles/model', as: :model
