+++ app/views/api/v1/fleet_vehicles/public.jbuilder
@@ -1,0 +1,3 @@
+# frozen_string_literal: true
+
+json.array! @vehicles, partial: 'api/v1/vehicles/minimal_public', as: :vehicle
