+++ app/views/api/v1/fleet_vehicles/public_models.jbuilder
@@ -1,0 +1,3 @@
+# frozen_string_literal: true
+
+json.array! @models, partial: 'api/v1/fleet_vehicles/public_model', as: :model
