+++ app/views/api/v1/fleet_vehicles/quick_stats.jbuilder
