--- app/views/api/v1/fleets/members.jbuilder
@@ -1,3 +1,0 @@
-# frozen_string_literal: true
-
-json.array! @members, partial: 'api/v1/fleets/member', as: :member
