--- app/views/api/v1/fleets/models.jbuilder
@@ -1,3 +1,0 @@
-# frozen_string_literal: true
-
-json.array! @models, partial: 'api/v1/fleets/model', as: :model
