+++ db/migrate/20210402203006_add_description_to_fleets.rb
@@ -1,0 +1,5 @@
+class AddDescriptionToFleets < ActiveRecord::Migration[6.1]
+  def change
+    add_column :fleets, :description, :text
+  end
+end
